% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-data-search.R
\name{addDataSearchMarker}
\alias{addDataSearchMarker}
\title{Add a data search marker control to the map.}
\usage{
addDataSearchMarker(map, group, position = "topright",
  propertyName = "label", initial = FALSE)
}
\arguments{
\item{map}{a map widget object}

\item{group}{a searchable group}

\item{position}{standard \href{http://leafletjs.com/reference.html#control-positions}{Leaflet control position options}.}

\item{propertyName}{property in marker.options(or feature.properties for
vector layer) trough filter elements in layer}

\item{initial}{search elements only by initial text}
}
\value{
modified map
}
\description{
Add a data search marker control to the map.
}
\examples{
data(quakes)
leaflet(data = quakes) \%>\%
    addTiles() \%>\%
    addMarkers(~long, ~lat, popup = ~as.character(mag),
               group = 'marker', label = ~as.character(mag)) \%>\%
    addDataSearchMarker('marker', position = 'topleft', propertyName = 'popup')
}

