% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-map-search.R
\name{leafletMapSearch}
\alias{leafletMapSearch}
\title{Add a Mapzen geocoder search control to the map.}
\usage{
leafletMapSearch(map)
}
\arguments{
\item{map}{a map widget object}
}
\description{
Add a Mapzen geocoder search control to the map.
}
\examples{
leaflet(data = quakes) \%>\%
addTiles() \%>\%
    leafletMapSearch()
}

